+++ docs/changelog.rst
@@ -6,6 +6,9 @@ This page details the changes in the various ``django-formtools`` releases.
 1.0 (2015-03-25)
 ----------------
 
+- Added the ``request`` parameter to :meth:`FormPreview.parse_params()
+  <formtools.preview.FormPreview.parse_params>`.
+
 - Added the ``request`` parameter to :meth:`WizardView.get_prefix()
   <formtools.wizard.views.WizardView.get_prefix>`.